Fix up o.e.snapshots split package in xpack-core
Moves the classes related to source-only snapshots to `o.e.s.sourceonly`
to avoid splitting the snapshots package across modules.
