+++ x-pack/plugin/core/build.gradle
@@ -148,10 +148,5 @@ testClusters.all {
 tasks.named('splitPackagesAudit').configure {
   // engine is owned by server, these should be renamed
   ignoreClasses 'org.elasticsearch.index.engine.FrozenEngine',
-    'org.elasticsearch.index.engine.RewriteCachingDirectoryReader',
-
-    // snapshots are part of server, these should be renamed
-    'org.elasticsearch.snapshots.SeqIdGeneratingFilterReader',
-    'org.elasticsearch.snapshots.SourceOnlySnapshot',
-    'org.elasticsearch.snapshots.SourceOnlySnapshotRepository'
+    'org.elasticsearch.index.engine.RewriteCachingDirectoryReader'
 }
